+++ src/ipc.c
@@ -675,7 +675,7 @@ static void ipc_handle(int sd)
 	close(client);
 }
 
-void ipc_init(char *sockfile)
+void ipc_init(char *sockfile, char *ident)
 {
 	socklen_t len;
 	int sd;
@@ -693,7 +693,7 @@ void ipc_init(char *sockfile)
 	if (sockfile)
 		strlcpy(sun.sun_path, sockfile, sizeof(sun.sun_path));
 	else
-		snprintf(sun.sun_path, sizeof(sun.sun_path), _PATH_MROUTED_SOCK);
+		snprintf(sun.sun_path, sizeof(sun.sun_path), _PATH_MROUTED_SOCK, ident);
 
 	unlink(sun.sun_path);
 	logit(LOG_DEBUG, 0, "Binding IPC socket to %s", sun.sun_path);